Change RequestId and ProgressToken to use int64_t instead of int (#121)
JSON-RPC allows string or number for request IDs. Using int64_t instead
of int prevents overflow when generating numeric request IDs from
uint64_t counters.

Changes:
- Update RequestId type from variant<std::string, int> to variant<std::string, int64_t>
- Update ProgressToken type to use int64_t for consistency
- Add factory function for int64_t request IDs
- Update all usages of holds_alternative<int> to holds_alternative<int64_t>
- Update all usages of get<int> to get<int64_t> for RequestId/ProgressToken
- Change pending_requests_ map key from int to int64_t in RequestTracker
- Cast uint64_t request ID counters to int64_t to avoid overflow

+++ include/mcp/client/mcp_client.h
@@ -293,15 +293,15 @@ class RequestTracker {
   // Add request to tracking
   void trackRequest(RequestPtr request) {
     std::lock_guard<std::mutex> lock(mutex_);
-    // Extract int ID from RequestId
-    int id = holds_alternative<int>(request->id) ? get<int>(request->id) : 0;
+    // Extract int64_t ID from RequestId
+    int64_t id = holds_alternative<int64_t>(request->id) ? get<int64_t>(request->id) : 0;
     pending_requests_[id] = request;
   }
 
   // Get request by ID without removing
   RequestPtr getRequest(const RequestId& id) {
     std::lock_guard<std::mutex> lock(mutex_);
-    int int_id = holds_alternative<int>(id) ? get<int>(id) : 0;
+    int64_t int_id = holds_alternative<int64_t>(id) ? get<int64_t>(id) : 0;
     auto it = pending_requests_.find(int_id);
     if (it != pending_requests_.end()) {
       return it->second;
@@ -312,7 +312,7 @@ class RequestTracker {
   // Remove and return request
   RequestPtr removeRequest(const RequestId& id) {
     std::lock_guard<std::mutex> lock(mutex_);
-    int int_id = holds_alternative<int>(id) ? get<int>(id) : 0;
+    int64_t int_id = holds_alternative<int64_t>(id) ? get<int64_t>(id) : 0;
     auto it = pending_requests_.find(int_id);
     if (it != pending_requests_.end()) {
       auto request = it->second;
@@ -336,7 +336,7 @@ class RequestTracker {
 
     // Remove timed out requests from tracking
     for (const auto& request : timed_out) {
-      int id = holds_alternative<int>(request->id) ? get<int>(request->id) : 0;
+      int64_t id = holds_alternative<int64_t>(request->id) ? get<int64_t>(request->id) : 0;
       pending_requests_.erase(id);
     }
 
@@ -352,7 +352,7 @@ class RequestTracker {
   mutable std::mutex mutex_;
   std::chrono::milliseconds timeout_;
   // Store using int ID extracted from RequestId
-  std::unordered_map<int, RequestPtr> pending_requests_;
+  std::unordered_map<int64_t, RequestPtr> pending_requests_;
 };
 
 /**
